Moved extension functions into different files. Event class now takes self-generic which caused huge changes.
