+++ gradle.properties
@@ -5,14 +5,14 @@ org.gradle.jvmargs=-Xmx4G
 # Fabric Properties
 # check these on https://fabricmc.net/develop
 minecraft_version=1.20.1
-loader_version=0.14.21
+loader_version=0.14.22
 kotlin_version=1.9.0
 ktfmt_version=0.12.0
 parchment_mappings=2023.07.30
 
 
 # Mod Properties
-mod_version = 0.1.5-beta
+mod_version = 0.2.0-beta
 maven_group = io.github.mosaicmc.mosaiccore
 mod_id = mosaiccore
