+++ src/main/kotlin/io/github/mosaicmc/mosaiccore/api/event/Listener.kt
@@ -18,8 +18,8 @@
 
 package io.github.mosaicmc.mosaiccore.api.event
 
-import io.github.mosaicmc.mosaiccore.api.plugin.PluginContainer
-import io.github.mosaicmc.mosaiccore.internal.event.ListenerImpl
+import io.github.mosaicmc.mosaiccore.api.event.subscriber.SubscriberData
+import io.github.mosaicmc.mosaiccore.api.event.subscriber.SubscriberFunction
 import kotlin.reflect.KClass
 
 /**
@@ -39,51 +39,9 @@ interface Listener {
      * @param data The subscriber data, which includes priority and cancellable settings (optional).
      * @param function The subscriber function, a lambda expression to handle the event.
      */
-    fun <E : Event> subscriber(
+    fun <E : Event<E>> subscriber(
         eventClazz: KClass<E>,
         data: SubscriberData = SubscriberData(),
-        function: SubscriberFunction<E>
+        function: SubscriberFunction<Event<E>>
     )
 }
-
-/**
- * DSL subscriber function
- *
- * This DSL function allows adding a subscriber to listen for events of type [E].
- *
- * @param E The event class that the subscriber will listen to.
- * @param data The subscriber data, which includes priority and cancellable settings (optional).
- * @param function The subscriber function, a lambda expression to handle the event.
- */
-inline fun <reified E : Event> Listener.subscriber(
-    data: SubscriberData = SubscriberData(),
-    noinline function: SubscriberFunction<E>
-) = subscriber(E::class, data, function)
-
-/**
- * DSL subscriber function with priority and cancellable settings
- *
- * This DSL function allows adding a subscriber to listen for events of type [E] with specific
- * priority and cancellable settings.
- *
- * @param E The event class that the subscriber will listen to.
- * @param priority The priority of the subscriber.
- * @param cancellable Determines whether canceled events should be ignored by the subscriber.
- * @param function The subscriber function, a lambda expression to handle the event.
- */
-inline fun <reified E : Event> Listener.subscriber(
-    priority: Priority = Priority.NORMAL,
-    cancellable: Boolean = false,
-    noinline function: SubscriberFunction<E>
-) = subscriber<E>(SubscriberData(priority, cancellable), function)
-
-/**
- * Listener DSL function
- *
- * This DSL function provides a convenient way to register multiple event subscribers.
- *
- * @param block The DSL block where event subscribers are defined.
- * @receiver The plugin container to which the listeners will be associated.
- */
-infix fun PluginContainer.listen(block: Listener.() -> Unit) =
-    ListenerImpl(this).apply(block).register()
