+++ src/main/kotlin/io/github/mosaicmc/mosaiccore/internal/CoreEvents.kt
