+++ src/main/resources/fabric.mod.json
@@ -19,7 +19,13 @@
     "preLaunch": [
       {
         "adapter": "kotlin",
-        "value": "io.github.mosaicmc.mosaiccore.internal.MainKt::preInit"
+        "value": "io.github.mosaicmc.mosaiccore.internal.Main::preInit"
+      }
+    ],
+    "plugin": [
+      {
+        "adapter": "kotlin",
+        "value": "io.github.mosaicmc.mosaiccore.internal.Main::test"
       }
     ]
   },
